pyproject.toml
[project]
name = "gmail-ingester"
version = "0.1.0"
description = "Fetch Gmail emails by label and convert HTML/text bodies to markdown"
readme = "README.md"
authors = [
    { name = "digster", email = "3542726+digster@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "trafilatura>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
gmail-ingester = "gmail_ingester.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
]
